+++ test/selfattentest.cpp
@@ -52,7 +52,7 @@ TEST(CommonTest, selfatten){
                         30, 31, 32, 33, 34, 35,
                         36, 37, 38, 39, 40, 41,
                         42, 43, 44, 45, 46, 47};
-    int mask[2];
+    uint64_t mask[2];
     mask[0] = 2;
     mask[1] = 4;
