+++ test/transformertest.cpp
@@ -150,7 +150,7 @@ TEST(CommonTest, transformer){
                         36, 37, 38, 39, 40, 41,
                         42, 43, 44, 45, 46, 47};
 
-    int mask[2] = {2, 4};
+    uint64_t mask[2] = {2, 4};
 
     float out[48];
